kata-agent: Check if attestation binaries are enabled as systemd units
In some environments the guest components, cdh, attestation-agent
etc are run as systemd units and this patch enables checking for systemd
enabled units before instantiating the processes

+++ src/agent/src/main.rs
@@ -37,6 +37,7 @@ use std::process::Command;
 use std::sync::Arc;
 use tracing::{instrument, span};
 
+
 mod cdh;
 mod config;
 mod console;
@@ -543,6 +544,29 @@ fn wait_for_path_to_exist(logger: &Logger, path: &str, timeout_secs: i32) -> Res
     Err(anyhow!("wait for {} to exist timeout.", path))
 }
 
+fn is_systemd_service_enabled(logger: &Logger, path: &str) -> Result<bool> {
+    let process_name = Path::new(path).file_name().unwrap().to_str().unwrap();
+    let service_name = format!("{}.service", process_name);
+
+    // Check if the systemd service is enabled
+    let status = Command::new("systemctl")
+        .args(&["is-enabled", &service_name])
+        .output()?;
+
+    let is_enabled = status.status.success();
+
+    // Log the path and the service_name
+    info!(logger, "Checking service"; "path" => path, "service_name" => &service_name);
+
+    if is_enabled {
+        info!(logger, "Service {} is enabled.", service_name);
+    } else {
+        info!(logger, "Service {} is not enabled.", service_name);
+    }
+
+    Ok(is_enabled)
+}
+
 fn launch_process(
     logger: &Logger,
     path: &str,
@@ -553,10 +577,23 @@ fn launch_process(
     if !Path::new(path).exists() {
         return Err(anyhow!("path {} does not exist.", path));
     }
+
+    // Check if the systemd service is enabled
+    if is_systemd_service_enabled(logger, path)? {
+        info!(logger, "Skipping launch for the process: {} as it's enabled via systemd", path);
+        if !unix_socket_path.is_empty() && timeout_secs > 0 {
+            wait_for_path_to_exist(logger, unix_socket_path, timeout_secs)?;
+        }
+        return Ok(());
+    }
+
+    // If the service is not enabled, proceed with launching it
     if !unix_socket_path.is_empty() && Path::new(unix_socket_path).exists() {
         fs::remove_file(unix_socket_path)?;
     }
+    
     Command::new(path).args(args).spawn()?;
+    
     if !unix_socket_path.is_empty() && timeout_secs > 0 {
         wait_for_path_to_exist(logger, unix_socket_path, timeout_secs)?;
     }
